Always use uint64_t for internal timestamp values
In most places we use 64 bit unsigned integers; lets be consistent and
use it everywhere.

+++ src/evdev.c
@@ -129,7 +129,7 @@ update_key_down_count(struct evdev_device *device, int code, int pressed)
 
 void
 evdev_keyboard_notify_key(struct evdev_device *device,
-			  uint32_t time,
+			  uint64_t time,
 			  int key,
 			  enum libinput_key_state state)
 {
@@ -144,7 +144,7 @@ evdev_keyboard_notify_key(struct evdev_device *device,
 
 void
 evdev_pointer_notify_physical_button(struct evdev_device *device,
-				     uint32_t time,
+				     uint64_t time,
 				     int button,
 				     enum libinput_button_state state)
 {
@@ -159,7 +159,7 @@ evdev_pointer_notify_physical_button(struct evdev_device *device,
 
 void
 evdev_pointer_notify_button(struct evdev_device *device,
-			    uint32_t time,
+			    uint64_t time,
 			    int button,
 			    enum libinput_button_state state)
 {
